# cursos/src/main/webapp/openapi/cursos.yaml
openapi: '3.0.2'
info:
  title: Cursos
  version: '1.0'
servers:
  - url: http://localhost:8080/cursos/api/v1

tags:
  - name: Reseñas
  - name: Cursos
paths:
  /resenyas:
    get:
      tags:
        - Reseñas
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resenya'
    post:
      tags:
        - Reseñas
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenya'
        required: true
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resenya'
    delete:
      tags:
        - Reseñas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: NO CONTENT
  /resenyas/{codigo}:
    get:
      tags:
      - Reseñas
      parameters:
      - name: codigo
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resenya'
    put:
      tags:
        - Reseñas
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenya'
        required: true
      responses:
        '200':
          description: OK
  /cursos:
    get:
      tags:
       - Cursos
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'
  /cursos/resenyas:
    get:
      tags:
       - Cursos
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'

  /cursos/{codigo}:
    get:
      tags:
      - Cursos
      parameters:
      - name: codigo
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'
  /cursos/r/{codigo}:
    get:
      tags:
      - Cursos
      parameters:
      - name: codigo
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
                schema:
                    type: array
                    items:
                      $ref: '#/components/schemas/Curso'

components:
  schemas:
    Resenya:
      type: object
      properties:
        alumno_codigo:
          type: integer
          format: int64
        curso_codigo:
          type: integer
          format: int64
        resenya:
          type: string
        resenyas_id:
          type: integer
          format: int64
    Curso:
      type: object
      properties:
        codigo:
          type: integer
          format: int64
        nombre:
          type: string
        identificador:
          type: string
      
        resenya:
          $ref: '#/components/schemas/Resenya'